@@ -11,23 +11,24 @@
         "php": "^8.2",
         "guzzlehttp/guzzle": "^7.9",
         "http-interop/http-factory-guzzle": "^1.0",
-        "laravel/framework": "^11.0",
-        "laravel/horizon": "^5.22",
-        "laravel/tinker": "^2.9",
-        "laravel/ui": "^4.4",
-        "mailerlite/laravel-elasticsearch": "^11",
         "meilisearch/meilisearch-php": "^1.10",
+        "laravel/framework": "^12.2",
+        "laravel/horizon": "^5.30",
+        "laravel/tinker": "^2.10.1",
+        "laravel/ui": "^4.6",
+        "mailerlite/laravel-elasticsearch": "^11.2",
         "mobiledetect/mobiledetectlib": "^4.8",
-        "nunomaduro/collision": "^8.0",
+        "nunomaduro/collision": "^8.6",
         "squizlabs/php_codesniffer": "^3.6"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.10",
+        "barryvdh/laravel-debugbar": "^3.15",
         "fakerphp/faker": "^1.23",
-        "friendsofphp/php-cs-fixer": "^3.64",
+        "friendsofphp/php-cs-fixer": "^3.66",
         "mockery/mockery": "^1.6",
-        "phpunit/phpunit": "^11.2",
-        "spatie/laravel-ignition": "^2.4"
+        "phpunit/phpunit": "^11.5.3",
+        "spatie/laravel-ignition": "^2.9",
+        "laravel/pail": "^1.2.2"
     },
     "config": {
         "optimize-autoloader": true,
@@ -59,7 +60,7 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "scripts": {
         "post-autoload-dump": [
@@ -71,6 +72,10 @@
         ],
         "post-create-project-cmd": [
             "@php artisan key:generate --ansi"
+        ],
+        "dev": [
+            "Composer\\Config::disableProcessTimeout",
+            "npx concurrently -c \"#93c5fd,#c4b5fd,#fb7185,#fdba74\" \"php artisan serve\" \"php artisan queue:listen --tries=1\" \"php artisan pail --timeout=0\" \"npm run dev\" --names=server,queue,logs,vite"
         ]
     }
 }
